+++ xinference/web/ui/src/scenes/launch_model/imageCard.js
