+++ xinference/web/ui/src/scenes/launch_model/index.js
@@ -70,8 +70,8 @@ const LaunchModel = () => {
             <Tab label="Language Models" value="1" />
             <Tab label="Embedding Models" value="2" />
             <Tab label="Rerank Models" value="3" />
-            <Tab label="Custom Models" value="4" />
-            <Tab label="Image Models" value="5" />
+            <Tab label="Image Models" value="4" />
+            <Tab label="Custom Models" value="5" />
           </TabList>
         </Box>
         <TabPanel value="1" sx={{ padding: 0 }}>
@@ -83,11 +83,11 @@ const LaunchModel = () => {
         <TabPanel value="3" sx={{ padding: 0 }}>
           <LaunchRerank />
         </TabPanel>
-        <TabPanel value="4" sx={{ padding: 0 }}>
-          <LaunchCustom gpuAvailable={gpuAvailable} />
-        </TabPanel>
-        <TabPanel value="5" sx={{ paddding: 0 }}>
+        <TabPanel value="4" sx={{ paddding: 0 }}>
           <LaunchImage />
+        </TabPanel>
+        <TabPanel value="5" sx={{ padding: 0 }}>
+          <LaunchCustom gpuAvailable={gpuAvailable} />
         </TabPanel>
       </TabContext>
     </Box>
